--- lib/discovery/v1/event.js
@@ -1,9 +1,0 @@
-import all from '../../utils/all';
-import find from '../../utils/find';
-
-import {api} from './';
-
-export default (apiKey) => ({
-  all: all(api, apiKey, 'events'),
-  find: find(api, apiKey, 'events')
-});
