+++ lib/discovery/v2.js
@@ -1,5 +1,5 @@
-import all from '../../utils/all';
-import find from '../../utils/find';
+import all from '../utils/all';
+import find from '../utils/find';
 
 const api = 'discovery/v2';
